refactor: :recycle: Separate request error handlers and error.
