+++ utils/requests/requestError.ts
@@ -1,0 +1,6 @@
+export type RequestError<T extends Record<string, string | undefined>> =
+  | BadRequestError<T>
+  | UnauthorizedError<Record<string, string>>
+  | ForbiddenError<Record<string, string>>
+  | NotFoundError
+  | ServerError;
