+++ utils/requests/requestErrorMessage.ts
@@ -1,0 +1,3 @@
+export type RequestErrorMessage<T> = Partial<
+  Record<keyof T, string | undefined>
+>;
