+++ utils/requests/requestResponse.ts
@@ -1,3 +1,6 @@
+import { RequestError } from "./requestError";
+import { RequestErrorMessage } from "./requestErrorMessage";
+
 export type RequestResponse<TData, TInput = Record<string, never>> =
   | TData
   | RequestError<RequestErrorMessage<TInput>>;
